+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Next version](https://github.com/sdss/sdsstools/compare/1.0.2...main)
 
 - [#35](https://github.com/sdss/sdsstools/issues/35) Add support to reload a `Configuration`. If the configuration is based on files on disk and those have changed, the files will be read again.
+- Add the option to print the time of a log message in `StreamFormatter` by doing `log.sh.formatter.print_time = True`.
 
 
 ## [1.0.2](https://github.com/sdss/sdsstools/compare/1.0.1...1.0.2)